agents/backend/onyx/server/features/agents/backend_ads/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.24.0
beautifulsoup4>=4.9.3
pydantic>=2.7.4
pydantic-settings>=2.4.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
requests>=2.26.0
lxml>=4.9.0
tenacity>=8.0.1
python-multipart>=0.0.5

# Dependencias de Langchain
langchain>=0.3.25
langchain-community>=0.3.25
langchain-core>=0.3.65

# Utilidades
cachetools # Para caching

# New additions
rembg>=2.0.50
onnxruntime>=1.15.0  # Required by rembg
pillow>=10.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
numpy>=1.24.0
opencv-python-headless>=4.7.0.72
scikit-image>=0.21.0
aiofiles>=23.1.0
tqdm>=4.65.0
loguru>=0.7.0
pyvips>=2.2.1
imageio>=2.31.0
gunicorn>=21.2.0
starlette>=0.27.0
psutil>=5.9.0
blosc>=1.11.1
redis>=5.0.1
redis[hiredis]>=5.0.1
aioredis==2.0.1